7616273330601201434/testbed/requirements.txt
langchain>=1.0.0
langgraph>=1.0.0
langchain-core>=1.0.0
langchain-openai>=0.2.0
langchain-community>=0.2.0
pydantic>=2.7.0
PyYAML>=6.0.1
python-dotenv>=1.0.1
faiss-cpu
sentence-transformers
pytest
chromadb
rank-bm25
